Move __archive_write_entry_filetype_unsupported into separate source file
__archive_write_entry_filetype_unsupported is called by several archive_write_set_format_XXX.
Using the static version of libarchive archive_set_format.o is always included into
the binary. Thus all other formats and the compression library dependencies are included as well.
This is because archive_set_format.c declares the code table with all functions.

This fix reduces the dependency to the compression formats and allows a user to use an
uncompressed format (e.g. archive_write_set_format_ustar) without depending on the compression
libraries.
